requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.14=pypi_0
aiosignal=1.4.0=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.9.0=pypi_0
argbind=0.3.9=pypi_0
asttokens=3.0.0=pyhd8ed1ab_1
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
audioread=3.0.1=pypi_0
azure-cognitiveservices-speech=1.45.0=pypi_0
azure-core=1.35.0=pypi_0
brotli-python=1.1.0=py310hf71b8c6_3
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.7.14=hbd8a1cb_0
cachetools=5.5.2=pypi_0
certifi=2025.7.14=pyhd8ed1ab_0
cffi=1.17.1=py310h8deb56e_0
charset-normalizer=3.4.2=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_1
comm=0.2.2=pyhd8ed1ab_1
contourpy=1.3.2=pypi_0
cpython=3.10.18=py310hd8ed1ab_0
crcmod=1.7=pypi_0
cryptography=45.0.5=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.8.15=py310h25320af_0
decorator=5.2.1=pyhd8ed1ab_0
descript-audio-codec=1.0.0=pypi_0
descript-audiotools=0.7.4=pypi_0
distro=1.9.0=pypi_0
docstring-parser=0.17.0=pypi_0
editdistance=0.8.1=pypi_0
einops=0.8.1=pypi_0
exceptiongroup=1.3.0=pyhd8ed1ab_0
executing=2.2.0=pyhd8ed1ab_0
expat=2.7.1=h6a678d5_0
ffmpy=0.6.1=pypi_0
filelock=3.18.0=pyhd8ed1ab_0
fire=0.7.0=pypi_0
flash-attn=2.8.2=pypi_0
flatten-dict=0.4.2=pypi_0
fonttools=4.59.0=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.7.0=pyhd8ed1ab_0
funasr=1.2.6=pypi_0
future=1.0.0=pypi_0
gmp=6.3.0=h6a678d5_0
gmpy2=2.2.1=py310he8512ff_0
google-auth=2.40.3=pypi_0
google-genai=1.27.0=pypi_0
h11=0.16.0=pypi_0
h2=4.2.0=pyhd8ed1ab_0
hf-xet=1.1.5=pypi_0
hpack=4.1.0=pyhd8ed1ab_0
httpcore=1.0.9=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.33.4=pypi_0
hydra-core=1.3.2=pypi_0
hyperframe=6.1.0=pyhd8ed1ab_0
idna=3.10=pyhd8ed1ab_1
importlib-metadata=8.7.0=pyhe01879c_1
importlib-resources=6.5.2=pypi_0
ipykernel=6.29.5=pyh3099207_0
ipython=8.37.0=pyh8f84b5b_0
jaconv=0.4.0=pypi_0
jamo=0.4.1=pypi_0
jedi=0.19.2=pyhd8ed1ab_1
jieba=0.42.1=pypi_0
jinja2=3.1.6=pyhd8ed1ab_0
jiter=0.10.0=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.5.1=pypi_0
julius=0.2.7=pypi_0
jupyter_client=8.6.3=pyhd8ed1ab_1
jupyter_core=5.8.1=pyh31011fe_0
kaldiio=2.18.1=pypi_0
kiwisolver=1.4.8=pypi_0
krb5=1.21.3=h8a1dbc1_1
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libabseil=20240116.2=cxx17_h6a678d5_0
libblas=3.9.0=32_h59b9bed_openblas
libcblas=3.9.0=32_he106b2a_openblas
libedit=3.1.20230828=h5eee18b_0
libffi=3.4.4=h6a678d5_1
libgcc=15.1.0=h767d61c_3
libgcc-ng=15.1.0=h69a702a_3
libgfortran=15.1.0=h69a702a_3
libgfortran5=15.1.0=hcea5267_3
libgomp=15.1.0=h767d61c_3
liblapack=3.9.0=32_h7ac8fdf_openblas
libopenblas=0.3.30=pthreads_h94d23a6_1
libprotobuf=4.25.3=he621ea3_0
librosa=0.11.0=pypi_0
libsodium=1.0.20=h4ab18f5_0
libstdcxx=15.1.0=h8f9b012_3
libstdcxx-ng=11.2.0=h1234567_1
libtorch=2.4.1=cpu_generic_hb3b73e9_0
libuuid=1.41.5=h5eee18b_0
libuv=1.51.0=hb9d3cd8_0
libxcb=1.17.0=h9b100fa_0
lightning=2.5.2=pyhd8ed1ab_0
lightning-utilities=0.14.3=pyhd8ed1ab_0
llvmlite=0.44.0=pypi_0
markdown-it-py=3.0.0=pypi_0
markdown2=2.5.3=pypi_0
markupsafe=3.0.2=py310h89163eb_1
matplotlib=3.10.3=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_1
mdurl=0.1.2=pypi_0
modelscope=1.28.1=pypi_0
mpc=1.3.1=h24ddda3_1
mpfr=4.2.1=h90cbb55_3
mpmath=1.3.0=pyhd8ed1ab_1
msgpack=1.1.1=pypi_0
multidict=6.6.3=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pyhd8ed1ab_1
networkx=3.4.2=pyh267e887_2
nomkl=1.0=h5ca1d4c_0
numba=0.61.2=pypi_0
numpy=1.26.4=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=9.1.0.70=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-cufile-cu12=1.11.1.6=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-cusparselt-cu12=0.6.3=pypi_0
nvidia-nccl-cu12=2.20.5=pypi_0
nvidia-nvjitlink-cu12=12.6.85=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
omegaconf=2.3.0=pypi_0
openai=1.97.1=pypi_0
openssl=3.5.1=h7b32b05_0
oss2=2.19.1=pypi_0
packaging=25.0=pyh29332c3_1
pandas=2.3.1=pypi_0
parler-tts=0.2.2=pypi_0
parso=0.8.4=pyhd8ed1ab_1
pexpect=4.9.0=pyhd8ed1ab_1
pickleshare=0.7.5=pyhd8ed1ab_1004
pillow=11.3.0=pypi_0
pip=25.1=pyhc872135_2
platformdirs=4.3.8=pyhe01879c_0
pooch=1.8.2=pypi_0
prompt-toolkit=3.0.51=pyha770c72_0
propcache=0.3.2=pypi_0
psutil=7.0.0=py310ha75aee5_0
pthread-stubs=0.3=h0ce48e5_1
ptyprocess=0.7.0=pyhd8ed1ab_1
pure_eval=0.2.3=pyhd8ed1ab_1
pyasn1=0.6.1=pypi_0
pyasn1-modules=0.4.2=pypi_0
pycparser=2.22=pyh29332c3_1
pycryptodome=3.23.0=pypi_0
pydantic=2.11.7=pypi_0
pydantic-core=2.33.2=pypi_0
pygments=2.19.2=pyhd8ed1ab_0
pyloudnorm=0.1.1=pypi_0
pynndescent=0.5.13=pypi_0
pyparsing=3.2.3=pypi_0
pysocks=1.7.1=pyha55dd90_7
pystoi=0.4.1=pypi_0
python=3.10.18=h1a3bd86_0
python-dateutil=2.9.0.post0=pyhe01879c_2
python_abi=3.10=2_cp310
pytorch-wpe=0.0.1=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.2=py310h89163eb_2
pyzmq=27.0.0=py310h71f11fc_0
randomname=0.2.1=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.4=pyhd8ed1ab_0
rich=14.0.0=pypi_0
rsa=4.9.1=pypi_0
safetensors=0.5.3=pypi_0
scikit-learn=1.7.1=pypi_0
scipy=1.15.3=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=78.1.1=py310h06a4308_0
six=1.17.0=pyhe01879c_1
sleef=3.8=h1b44611_0
sniffio=1.3.1=pypi_0
soundfile=0.13.1=pypi_0
soxr=0.5.0.post1=pypi_0
sqlite=3.50.2=hb25bd0a_1
stack_data=0.6.3=pyhd8ed1ab_1
sympy=1.14.0=pyh2585a3b_105
tenacity=8.5.0=pypi_0
tensorboardx=2.6.4=pypi_0
termcolor=3.1.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h993c535_1
tokenizers=0.20.3=pypi_0
torch=2.4.1+cu121=pypi_0
torch-complex=0.4.4=pypi_0
torch-stoi=0.2.3=pypi_0
torchaudio=2.4.1+cu121=pypi_0
torchmetrics=1.7.4=pyhd8ed1ab_0
tornado=6.5.1=py310ha75aee5_0
tqdm=4.67.1=pyhd8ed1ab_1
traitlets=5.14.3=pyhd8ed1ab_1
transformers=4.46.1=pypi_0
triton=3.0.0=pypi_0
typing-extensions=4.14.1=h4440ef1_0
typing-inspection=0.4.1=pypi_0
typing_extensions=4.14.1=pyhe01879c_0
tzdata=2025b=h04d1e81_0
umap-learn=0.5.9.post2=pypi_0
urllib3=2.5.0=pyhd8ed1ab_0
wcwidth=0.2.13=pyhd8ed1ab_1
websockets=15.0.1=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
yaml=0.2.5=h7f98852_2
yarl=1.20.1=pypi_0
zeromq=4.3.5=h3b0a872_7
zipp=3.23.0=pyhd8ed1ab_0
zlib=1.2.13=h5eee18b_1
zstandard=0.23.0=py310ha75aee5_2
